+++ sources/ANSW2016/create-data_dictionary_ANSW1016.R
@@ -1,8 +1,5 @@
-### Creación de dos diccionarios ES para quanteda ###
-### Creation of two dictionaries ES (Spanish) for quanteda  ###
-
-### Uno de valencias (Hinojosa et al. 2016) y otro de polaridad (Cruz et al. 2014)
-### One for valences (Hinojosa et al. 2016) and another for polarity (Cruz et al. 2014)
+### Creación de un diccionarios ES de valencias para quanteda usando datos de Hinojosa et al. 2016
+### Creation of an ES valence dictionary for quanteda using data from Hinojosa et al. 2016
 
 ### Contact: mariano.rico@upm.es
 
@@ -27,47 +24,47 @@ df1 <- read.xlsx(file = "Hinojosa2016.xlsx",
 #Primero, hago el dict con las palabras
 #First, I create the dict with the words
 library(quanteda) #To know what a dictionary is
-data_dictionary_ANSW <- dictionary(list(valence = df1$Word))
+data_dictionary_ANSW2016 <- dictionary(list(valence = df1$Word))
 #Ahora le asigno las valencias
 #Now I assign the valences
 library(quanteda.sentiment) #To know what is valence()
-valence(data_dictionary_ANSW) <- list(valence = df1$Val_Mn)
+valence(data_dictionary_ANSW2016) <- list(valence = df1$Val_Mn)
 
 #Copio, cambio nombre y asigno
 #Copy, change name and assign
 #Arousal
-data_dictionary_ANSW["arousal"] <- data_dictionary_ANSW["valence"]
-valence(data_dictionary_ANSW)["arousal"] <- list(arousal = df1$Ar_Mn)
+data_dictionary_ANSW2016["arousal"] <- data_dictionary_ANSW2016["valence"]
+valence(data_dictionary_ANSW2016)["arousal"] <- list(arousal = df1$Ar_Mn)
 
 #Happiness
-data_dictionary_ANSW["happiness"] <- data_dictionary_ANSW["valence"]
-valence(data_dictionary_ANSW)["happiness"] <- list(happiness = df1$Hap_Mn)
+data_dictionary_ANSW2016["happiness"] <- data_dictionary_ANSW2016["valence"]
+valence(data_dictionary_ANSW2016)["happiness"] <- list(happiness = df1$Hap_Mn)
 
 #Anger
-data_dictionary_ANSW["anger"] <- data_dictionary_ANSW["valence"]
-valence(data_dictionary_ANSW)["anger"] <- list(anger = df1$Ang_Mn)
+data_dictionary_ANSW2016["anger"] <- data_dictionary_ANSW2016["valence"]
+valence(data_dictionary_ANSW2016)["anger"] <- list(anger = df1$Ang_Mn)
 
 #Sadness
-data_dictionary_ANSW["sadness"] <- data_dictionary_ANSW["valence"]
-valence(data_dictionary_ANSW)["sadness"] <- list(sadness = df1$Sad_Mn)
+data_dictionary_ANSW2016["sadness"] <- data_dictionary_ANSW2016["valence"]
+valence(data_dictionary_ANSW2016)["sadness"] <- list(sadness = df1$Sad_Mn)
 
 #Fearness
-data_dictionary_ANSW["fearness"] <- data_dictionary_ANSW["valence"]
-valence(data_dictionary_ANSW)["fearness"] <- list(fearness = df1$Fear_Mn)
+data_dictionary_ANSW2016["fearness"] <- data_dictionary_ANSW2016["valence"]
+valence(data_dictionary_ANSW2016)["fearness"] <- list(fearness = df1$Fear_Mn)
 
 #Disgust
-data_dictionary_ANSW["disgust"] <- data_dictionary_ANSW["valence"]
-valence(data_dictionary_ANSW)["disgust"] <- list(disgust = df1$Disg_Mn)
+data_dictionary_ANSW2016["disgust"] <- data_dictionary_ANSW2016["valence"]
+valence(data_dictionary_ANSW2016)["disgust"] <- list(disgust = df1$Disg_Mn)
 
 #Concreteness
-data_dictionary_ANSW["concreteness"] <- data_dictionary_ANSW["valence"]
-valence(data_dictionary_ANSW)["concreteness"] <- list(concreteness = df1$Con_Mn)
+data_dictionary_ANSW2016["concreteness"] <- data_dictionary_ANSW2016["valence"]
+valence(data_dictionary_ANSW2016)["concreteness"] <- list(concreteness = df1$Con_Mn)
 
-meta(data_dictionary_ANSW) <- list(
+meta(data_dictionary_ANSW2016) <- list(
   title = "Dictionary created from Hinojosa, J. A. et al. (2016) 'Affective norms of 875 Spanish words for five discrete emotional categories and two emotional dimensions'.",
   description = "This diccionary has values for valence, arousal, concreteness, as well as five emotions: happiness, anger, sadness, fearness and disgust.",
   url = "https://link.springer.com/article/10.3758/s13428-015-0572-5",
   reference = "Hinojosa, J.A., Martínez-García, N., Villalba-García, C. et al. Affective norms of 875 Spanish words for five discrete emotional categories and two emotional dimensions. Behav Res 48, 272–284 (2016)",
   license = "unknonwn"
 )
-usethis::use_data(data_dictionary_ANSW, overwrite = TRUE)
+usethis::use_data(data_dictionary_ANSW2016, overwrite = TRUE)